Decrease timeout for fsharplex test cases.
The timeout was originally set to two minutes because certain input specifications (e.g., the fsharp lexer) hit a performance edge case that greatly slowed things down. That was fixed a while back; now, even for large input specifications fsharplex should finish running in a few seconds or less (on a fast machine).